重新整理项目结构，采用多文件注册指令
