+++ src/__init__.py
@@ -1,0 +1,4 @@
+from .chatbot import chatbot_reply
+from . import chatbot
+
+__all__ = ["chatbot_reply", "chatbot"]
