+++ src/bot/assistant.py
