+++ src/msg_const.py
@@ -1,0 +1,47 @@
+WELCOME_MSG = [
+    {
+        "type": "card",
+        "theme": "secondary",
+        "size": "lg",
+        "modules": [
+        {
+            "type": "header",
+            "text": {
+            "type": "plain-text",
+            "content": "欢迎加入CUFER'S HUB！"
+            }
+        },
+        {
+            "type": "section",
+            "text": {
+            "type": "kmarkdown",
+            "content": '''为了确保我们的社区安全，我们需要验证您的学生身份: 
+
+请私信我指令`/verif [学号]`进行身份验证。稍后您将会在学校邮箱中收到验证码。放心，您的学号经过加密处理，即使是管理员也无法查看您的学号。
+然后请用指令`/captcha [验证码]`进行验证。
+一定要注意输入格式哦！**不需要输入方括号**!格式错误将不会有任何响应！
+如果您已毕业，请您私信管理员进行验证。
+
+示例:
+```
+/verif 20xxxxxxxx
+/captcha 123456
+````
+
+**问：“诶？学校邮箱是什么？”**
+答：如果您没有申请过学校邮箱，请查阅[学校邮箱使用攻略](http://zhxy.cufe.edu.cn/info/1082/2097.htm)。申请成功之后再进行学号验证。
+**问：“怎么一直不发验证码？这bot是出问题了吗？”**
+答：请检查您的输入，确保您按照正确的指令格式书写，并且指令拼写无误，斜杠方向无误。如果仍然不行，请联系管理员。
+
+通过验证后，你将获得访问所有频道的权限。
+
+感谢您的配合，祝您在CUFER'S HUB愉快！
+
+如有任何问题，请联系管理员。
+            '''
+            }
+        }
+        ]
+    }
+]
+
